Move SingleLineInput to shared location
