+++ assets/blocks/quiz/answer-blocks/multiple-choice-answer-option.js
@@ -1,7 +1,7 @@
 import { Button } from '@wordpress/components';
 import { useEffect, useRef } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
-import SingleLineInput from '../../course-outline/single-line-input';
+import SingleLineInput from '../../../shared/blocks/single-line-input';
 import { OptionToggle } from './option-toggle';
 
 /**
