+++ assets/blocks/shared-style-editor.scss
@@ -1,2 +1,2 @@
 @import './button/button-editor';
-@import './course-outline/single-line-input/single-line-input';
+@import '../shared/blocks/single-line-input/single-line-input';
