+++ assets/shared/blocks/single-line-input/index.js
@@ -1,0 +1,3 @@
+import SingleLineInput from './single-line-input';
+
+export default SingleLineInput;
